houston_django/requirements.txt
boto3==1.34.108
celery==5.4.0
dj-database-url==2.1.0
django-ace==1.32.4
django-admin-autocomplete-filter==0.7.1
django-admin-interface==0.28.6
django-apptemplates==1.5
django-clearcache==1.2.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.2
django-phonenumber-field==7.3.0
django-storages==1.14.3
django==5.0.6
djangorestframework-simplejwt==5.3.1
djangorestframework==3.15.1
gunicorn==22.0.0
phonenumbers==8.13.37
psycopg2-binary==2.9.9
redis==5.0.4
requests==2.31.0
sentry-sdk==2.2.0
twilio==9.0.5
whitenoise==6.6.0
